Improve and change general design
