+++ app/src/main/res/drawable/arrow_background_round.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <solid android:color="@color/sub_purple" />
+    <corners
+        android:bottomRightRadius="24dp"
+        android:topRightRadius="24dp"
+        android:bottomLeftRadius="24dp"
+        android:topLeftRadius="24dp"
+        />
+</shape>
