+++ app/src/main/res/layout/card_widget_slim.xml
@@ -3,7 +3,7 @@
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
-        android:layout_height="48dp"
+        android:layout_height="@dimen/card_height"
         android:layout_marginBottom="10dp"
         android:background="@android:color/transparent">
