+++ app/src/main/res/layout/dialog_new_project.xml
@@ -50,7 +50,7 @@
                 app:layout_constraintTop_toBottomOf="@+id/editTextNewProjectName"/>
 
         <com.google.android.material.chip.ChipGroup
-                android:layout_width="wrap_content"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
                 android:animateLayoutChanges="true"
                 app:singleLine="false"
@@ -59,9 +59,10 @@
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent" app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/textView2" android:layout_marginTop="8dp"
-                android:layout_marginBottom="16dp" android:layout_marginEnd="@dimen/card_margin"
-                android:layout_marginStart="@dimen/card_margin"
-                android:id="@+id/langSelectionGroup">
+                android:layout_marginBottom="16dp" android:layout_marginEnd="8dp"
+                android:layout_marginStart="8dp"
+                android:id="@+id/langSelectionGroup" android:clipToPadding="true" android:paddingStart="16dp"
+                android:paddingEnd="16dp">
 
         </com.google.android.material.chip.ChipGroup>
