+++ app/src/main/res/layout/fragment_home.xml
@@ -8,9 +8,9 @@
                                                    tools:context=".ui.home.HomeFragment">
     <androidx.recyclerview.widget.RecyclerView
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
+            android:layout_height="wrap_content"
             app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent" app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginStart="8dp" android:layout_marginEnd="8dp"
+            android:layout_marginStart="16dp" android:layout_marginEnd="16dp"
             android:id="@+id/widgets"/>
 </androidx.constraintlayout.widget.ConstraintLayout>
