+++ app/src/main/res/layout/fragment_plugins.xml
@@ -27,7 +27,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:background="@drawable/gradient_card"
+            android:background="@color/main_purple"
             android:gravity="center_vertical">
 
             <ImageView
@@ -66,10 +66,10 @@
     <androidx.cardview.widget.CardView
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_marginStart="@dimen/card_margin"
-            android:layout_marginTop="20dp"
-            android:layout_marginEnd="@dimen/card_margin"
-            android:layout_marginBottom="@dimen/card_margin"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="16dp"
             app:cardCornerRadius="@dimen/card_radius"
             app:cardElevation="@dimen/card_elevation">
 
@@ -87,6 +87,8 @@
                 android:id="@+id/recyclerViewProjectList"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:clipToPadding="false" android:paddingTop="16dp"/>
+                android:clipToPadding="false" android:paddingTop="16dp"
+                android:paddingBottom="8dp"
+                android:paddingStart="8dp" android:paddingEnd="8dp"/>
     </androidx.cardview.widget.CardView>
 </LinearLayout>
