+++ app/src/main/res/layout/fragment_projects.xml
@@ -13,69 +13,87 @@
     app:layout_constraintTop_toTopOf="parent"
     tools:context=".ui.projects.ProjectsFragment">
 
-    <androidx.cardview.widget.CardView
-        android:id="@+id/cardViewProjectAlign"
-        android:layout_width="match_parent"
-        android:layout_height="@dimen/card_height"
-        android:layout_marginStart="@dimen/card_margin"
-        android:layout_marginEnd="@dimen/card_margin"
-        app:cardCornerRadius="@dimen/card_radius"
-        app:cardElevation="0dp"
-        app:cardMaxElevation="0dp">
-
-        <androidx.constraintlayout.widget.ConstraintLayout
+    <LinearLayout
+            android:orientation="horizontal"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/gradient_card">
+            android:layout_height="wrap_content">
+        <androidx.cardview.widget.CardView
+                android:id="@+id/cardViewProjectAlign"
+                android:layout_width="308dp"
+                android:layout_height="@dimen/card_height"
+                android:layout_marginStart="@dimen/card_margin"
+                app:cardCornerRadius="@dimen/card_radius"
+                app:cardElevation="0dp"
+                app:cardMaxElevation="0dp">
 
-            <ImageView
-                    android:id="@+id/alignStateIcon"
-                    android:layout_width="0dp"
-                    android:layout_height="20dp"
-                    android:src="@drawable/ic_round_sort_by_alpha_24"
-                    app:tint="#FFFFFF" app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
-                    android:layout_marginStart="20dp"/>
-            <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/nanumsquare_round_extrabold"
-                    android:text="@string/project_align_state"
-                    android:textColor="#FFFFFF"
-                    android:textSize="16sp"
-                    app:layout_constraintStart_toEndOf="@+id/alignStateIcon"
-                    app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintTop_toTopOf="parent"
-            />
-            <TextView
-                    android:id="@+id/alignState"
-                    android:layout_width="0dp"
-                    android:layout_height="wrap_content"
-                    android:layout_weight="0.15"
-                    android:fontFamily="@font/nanumsquare_round_extrabold"
-                    android:text="@string/default_placeholder"
-                    android:textColor="#FFFFFF"
-                    android:textSize="16sp" app:layout_constraintBottom_toBottomOf="parent"
-                    app:layout_constraintEnd_toEndOf="parent" app:layout_constraintTop_toTopOf="parent"
-                    android:layout_marginEnd="20dp"/>
-        </androidx.constraintlayout.widget.ConstraintLayout>
-    </androidx.cardview.widget.CardView>
+            <androidx.constraintlayout.widget.ConstraintLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:background="@color/main_purple">
+
+                <ImageView
+                        android:id="@+id/alignStateIcon"
+                        android:layout_width="0dp"
+                        android:layout_height="20dp"
+                        android:src="@drawable/ic_round_sort_by_alpha_24"
+                        app:tint="#FFFFFF" app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
+                        android:layout_marginStart="20dp"/>
+                <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:fontFamily="@font/nanumsquare_round_extrabold"
+                        android:text="@string/project_align_state"
+                        android:textColor="#FFFFFF"
+                        android:textSize="16sp"
+                        app:layout_constraintStart_toEndOf="@+id/alignStateIcon"
+                        app:layout_constraintBottom_toBottomOf="parent" app:layout_constraintTop_toTopOf="parent"
+                />
+                <TextView
+                        android:id="@+id/alignState"
+                        android:layout_width="0dp"
+                        android:layout_height="wrap_content"
+                        android:layout_weight="0.15"
+                        android:fontFamily="@font/nanumsquare_round_extrabold"
+                        android:text="@string/default_placeholder"
+                        android:textColor="#FFFFFF"
+                        android:textSize="16sp" app:layout_constraintBottom_toBottomOf="parent"
+                        app:layout_constraintEnd_toEndOf="parent" app:layout_constraintTop_toTopOf="parent"
+                        android:layout_marginEnd="20dp"/>
+            </androidx.constraintlayout.widget.ConstraintLayout>
+        </androidx.cardview.widget.CardView>
+
+        <com.google.android.material.floatingactionbutton.FloatingActionButton
+                android:id="@+id/fabNewProject"
+                android:layout_width="54dp"
+                android:layout_height="54dp"
+                android:layout_gravity="bottom|center_horizontal"
+                android:foregroundGravity="center"
+                android:padding="16dp"
+                android:src="@drawable/ic_round_add_white_24"
+                app:tint="@color/white" app:elevation="0dp"
+                app:rippleColor="@color/main_purple" app:backgroundTint="@color/main_purple"
+                android:layout_marginStart="8dp"/>
+    </LinearLayout>
 
     <androidx.cardview.widget.CardView
             android:id="@+id/cardProjectList"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginStart="@dimen/card_margin"
-            android:layout_marginTop="16dp"
-            android:layout_marginEnd="@dimen/card_margin"
-            android:layout_marginBottom="@dimen/card_margin"
-            app:cardBackgroundColor="@color/white"
-            app:cardCornerRadius="@dimen/card_radius"
-            app:cardElevation="@dimen/card_elevation">
+            android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="16dp"
+            android:layout_marginBottom="16dp"
+            app:cardCornerRadius="30dp"
+            app:cardElevation="@dimen/card_elevation" android:background="@android:color/white"
+            app:cardBackgroundColor="@android:color/white">
 
         <androidx.recyclerview.widget.RecyclerView
                 android:id="@+id/recyclerViewProjectList"
                 android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:clipToPadding="false" android:paddingTop="16dp" android:paddingStart="16dp"/>
+                android:layout_height="wrap_content"
+                android:clipToPadding="false" android:paddingTop="16dp"
+                android:paddingBottom="8dp"
+                android:paddingStart="8dp" android:paddingEnd="8dp"/>
     </androidx.cardview.widget.CardView>
 </LinearLayout>
