+++ app/src/main/res/layout/widget_uptime_default.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        android:background="@drawable/gradient_card"
+        android:background="@color/main_purple"
         android:layout_width="match_parent"
         android:layout_height="96dp">
