+++ app/src/main/res/layout/widget_uptime_slim.xml
@@ -2,10 +2,9 @@
 <androidx.constraintlayout.widget.ConstraintLayout
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
-        android:background="@drawable/gradient_card"
-        android:layout_height="48dp">
+        android:background="@color/main_purple"
+        android:layout_height="match_parent">
 
     <TextSwitcher
             android:id="@+id/uptimeText"
