+++ app/src/main/res/values/colors.xml
@@ -6,6 +6,9 @@
     <color name="teal_200">#FF03DAC5</color>
     <color name="teal_700">#FF018786</color>
 
+    <color name="main_purple">#ab9df2</color>
+    <color name="sub_purple">#F3F0FF</color>
+
     <color name="background">#F2F2F2</color>
     <color name="cardview">#FFFFFF</color>
     <color name="bottom_sheet">#74FFFFFF</color>
