+++ app/src/main/res/values/dimens.xml
@@ -8,10 +8,10 @@
     <dimen name="icon_margin">15dp</dimen>
     <dimen name="icon_width">24dp</dimen>
 
-    <dimen name="card_height">48dp</dimen>
-    <dimen name="card_radius">24dp</dimen>
+    <dimen name="card_height">54dp</dimen>
+    <dimen name="card_radius">27dp</dimen>
     <dimen name="card_elevation">0dp</dimen>
-    <dimen name="card_margin">8dp</dimen>
+    <dimen name="card_margin">20dp</dimen>
 
     <dimen name="default_card_height">50dp</dimen>
     <dimen name="default_vertical_padding">10dp</dimen>
